Resume analyzer/requirements.txt
pdf2image
pdfplumber
pytesseract
pyhton-dotenv
streamlit
google.generativeai